requirements.txt
opencv-python-headless==4.5.5.64
numpy==1.21.6
scikit-learn==1.0.2
tensorflow==2.8.0
Pillow==9.1.1
pytest==7.1.2  # For testing